+++ src/main/java/com/example/officebookingsystem/domain/implementation/UserDetailImpl.java
@@ -1,4 +1,4 @@
-package com.example.officebookingsystem.security.service;
+package com.example.officebookingsystem.domain.implementation;
 
 import com.example.officebookingsystem.domain.entity.User;
 import com.fasterxml.jackson.annotation.JsonIgnore;
