+++ src/main/java/com/example/officebookingsystem/service/AuthSerice.java
@@ -1,0 +1,4 @@
+package com.example.officebookingsystem.service;
+
+public class AuthSerice {
+}
